requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
openai
pandas
matplotlib
seaborn
fuzzywuzzy
python-Levenshtein
easyocr
cryptography
SpeechRecognition
audio-recorder-streamlit
numpy
oauth2client
xlsxwriter
sentence-transformers
torch
